requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.8.1
build==1.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
duckdb==0.10.0
fastapi==0.110.0
fsspec==2024.2.0
h11==0.14.0
htmltools==0.5.1
httpcore==1.0.4
httpx==0.27.0
humanize==4.9.0
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.3.0
Jinja2==3.1.3
joblib==1.3.2
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.0
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
palmerpenguins==0.1.4
pandas==2.2.1
pins==0.8.4
pip-tools==7.4.1
plotly==5.19.0
prompt-toolkit==3.0.36
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
requests==2.31.0
rsconnect_python==1.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==2.13.0
shiny==0.8.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tenacity==8.2.3
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.28.0
vetiver==0.2.4
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
xxhash==3.4.1
zipp==3.17.0
